backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15.0
drf-yasg>=1.21.7
djangorestframework-simplejwt>=5.3.1
psycopg2-binary>=2.9.11
django-filter>=2.4.0